# backend/services/merchant-service/api-docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "支付平台商户管理服务API文档",
        "title": "Merchant Service API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@payment-platform.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8002",
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/webhooks": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Webhook"
                ],
                "summary": "获取Webhook配置",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Webhook"
                ],
                "summary": "更新Webhook配置",
                "parameters": [
                    {
                        "description": "Webhook配置",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/payment-platform_merchant-service_internal_service.UpdateWebhookInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Webhook"
                ],
                "summary": "创建Webhook配置",
                "parameters": [
                    {
                        "description": "Webhook配置",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/payment-platform_merchant-service_internal_service.CreateWebhookInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Webhook"
                ],
                "summary": "删除Webhook配置",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "payment-platform_merchant-service_internal_service.CreateWebhookInput": {
            "type": "object",
            "required": [
                "events",
                "merchant_id",
                "url"
            ],
            "properties": {
                "events": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "type": "string"
                    }
                },
                "max_retries": {
                    "type": "integer"
                },
                "merchant_id": {
                    "type": "string"
                },
                "timeout_seconds": {
                    "type": "integer"
                },
                "url": {
                    "type": "string"
                }
            }
        },
...
        "payment-platform_merchant-service_internal_service.UpdateWebhookInput": {
            "type": "object",
            "properties": {
                "events": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "type": "string"
                    }
                },
                "is_enabled": {
                    "type": "boolean"
                },
                "max_retries": {
                    "type": "integer"
                },
                "timeout_seconds": {
                    "type": "integer"
                },
                "url": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}